+++ Cortex/Medable.framework/Headers/MDObjectInstance.h
@@ -10,6 +10,8 @@
 #import "MDBaseInstance.h"
 #import "MDObjectDefinition.h"
 
+NS_ASSUME_NONNULL_BEGIN
+
 /**
  * An object instance is a concrete object of a given class, also known as a context.
  * Each instance follows the form given by the Definition of that class.
@@ -26,7 +28,7 @@
 /**
  * The account id of the context updater.
  */
-@property (nonatomic, readonly) MDReference* updater;
+@property (nullable, nonatomic, readonly) MDReference* updater;
 
 /**
  * Object definition for this instance. Note that the definition class is a subinstance of the base definition.
@@ -38,16 +40,15 @@
  */
 @property (nonatomic, readonly) NSNumber* favorite;
 
-
-/**
- * Version number of the context.
- */
-@property (nonatomic, readonly) NSUInteger contextVersion;
-
 /**
  * Creates an object instance form the attributes provided.
  */
-+ (instancetype)objectInstanceWithAttributes:(NSDictionary*)attributes;
++ (nullable instancetype)objectInstanceWithAttributes:(NSDictionary*)attributes;
+
+/**
+ * Created object intstances from an array of attributes.
+ */
++ (nullable NSArray*)objectInstancesWithAttributes:(NSArray<NSDictionary *>*)attributes;
 
 // unavailable
 + (instancetype)new NS_UNAVAILABLE;
@@ -62,7 +63,7 @@
  * @param parameters Extra parameters in the API call.
  * @param callback The asynchronous callback, called upon receiving a response from the backend. If no MDFault is received in the callback, the synchronized connections are availble through the `connections` property.
  */
-- (void)synchronizeConnectionsWithParameters:(MDAPIParameters *)parameters
+- (void)synchronizeConnectionsWithParameters:(nullable MDAPIParameters *)parameters
                                     callback:(MDFaultCallback)callback;
 
 /**
@@ -72,7 +73,9 @@
  * @param parameters Extra parameters in the API call.
  * @param callback The asynchronous callback, called upon receiving a response from the backend. If no MDFault is received in the callback, the synchronized connections are availble through the `posts` property.
  */
-- (void)synchronizePostsWithParameters:(MDAPIParameters *)parameters
+- (void)synchronizePostsWithParameters:(nullable MDAPIParameters *)parameters
                               callback:(MDFaultCallback)callback;
 
 @end
+
+NS_ASSUME_NONNULL_END